[RUTA] Paso 4: Migrar blueprint entidades a app/modules/ (#93 Fase 3)
- app/modules/entidades/routes.py: Blueprint completo con template_folder='templates'
- Templates movidos (git mv): index, nueva, detalle
  de app/templates/entidades/ a app/modules/entidades/templates/entidades/
- app/modules/entidades/metadata.json: corregida ruta navigation (entidades.index)
- app/__init__.py: eliminado registro legacy de entidades.bp desde app/routes/
- app/routes/entidades.py: eliminado (migrado al módulo)
