+++ app/__init__.py
@@ -55,14 +55,13 @@ def create_app(config_name='development'):
         return Usuario.query.get(int(user_id))
 
     # Registrar blueprints - Rutas principales
-    from app.routes import auth, dashboard, proyectos, usuarios, perfil, entidades
+    from app.routes import auth, dashboard, proyectos, usuarios, perfil
 
     app.register_blueprint(auth.bp)
     app.register_blueprint(dashboard.bp)
     app.register_blueprint(proyectos.bp)
     app.register_blueprint(usuarios.bp)
     app.register_blueprint(perfil.bp)
-    app.register_blueprint(entidades.bp)        # Issue #61
 
     # Registrar blueprints - Wizards
     from app.routes import wizard_expediente
