+++ app/modules/entidades/templates/entidades/detalle.html
